test:unit testing for create_client and get_credit_points
