+++ Cargo.toml
@@ -15,3 +15,6 @@ rand = "0.8"
 [build-dependencies]
 tonic-build = "0.11"
 prost-build = "0.12"
+
+[dev-dependencies]
+tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
