+++ src/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod client_tests;
+pub mod credits_tests;
